pdfbot/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

chromadb==0.4.18
sentence-transformers==2.2.2

groq==0.4.2
pypdf==3.17.4
tiktoken==0.5.1

python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0